src/test/java/com/sy/miniproject/HelloControllerTest.java
package com.sy.miniproject;

import com.sy.miniproject.web.HelloController;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@WebMvcTest(controllers = HelloController.class)
public class HelloControllerTest {
    @Autowired
    private MockMvc mvc;

    @Test
    public void hello() throws Exception {
        String hello = "hello";

        mvc.perform(get("/hello"))
                .andExpect(status().isOk())
                .andExpect(content().string(hello)); // 결과검증
    }

    @Test
    public void helloDto() throws Exception {
        String name = "soyeon";
        int amount = 1000;

        mvc.perform(get("/hello/dto")
                .param("name", name)
                .param("amount", String.valueOf(amount))) // param은 String타입만 허용
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.name", is(name))) // jsonPath : json 응답값을 필드별로 검증할 수 있다.
                .andExpect(jsonPath("$.amount", is(amount)));

    }
}
